third_party/WebKit/LayoutTests/http/tests/loading/preload-img-test-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check http://xhr.spec.whatwg.org/.
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame

This test is to test whether preload image will require 2 more requests or not.

PASSED
